+++ include/item_menu.h
@@ -59,7 +59,7 @@ struct BagMenu
 {
     void (*newScreenCallback)(void);
     u8 tilemapBuffer[BG_SCREEN_SIZE];
-    u8 spriteIds[ITEMMENUSPRITE_COUNT];
+    u32 spriteIds[ITEMMENUSPRITE_COUNT];
     u8 windowIds[ITEMWIN_COUNT];
     u8 toSwapPos;
     u8 pocketSwitchDisabled:4;
